Add demo web endpoints
First, let's create a few web endpoints to understand how reactive types
can be used in a web application:

1. Synchronous return types can be still used and will be wrapped by the
framework. This is allowed as long as the controller method is not
blocking.

2. With the use of `Mono.never()`, we can see that the reactive type not
only controls the HTTP response body, but also when that response is
written.

3. The HTTP contract is not based on Servlet here, but rather on the
`ServerWebExchange` contract and `DataBuffer`.

4. Errors can also flow through the reactive pipeline and those should
be handled by application code or by the framework itself.